ScribeNote/requirements.txt
PyQt5==5.15.9
cryptography==41.0.3
pillow==10.0.0
